:fire: Remove test_patch_classifier.py and move the code to test_patch_predictor.py
